Refactor package for axelar vald heartbeats
