+++ docker/cvms/support_chains.yaml
@@ -139,6 +139,7 @@ axelar-dojo-1:
     - axelar-evm
     - voteindexer
     - axelar-amplifier-verifier
+    - axelar-vald-heartbeats
 axelar-testnet-lisbon-3:
   chain_name: axelar
   protocol_type: cosmos
@@ -153,6 +154,7 @@ axelar-testnet-lisbon-3:
     - axelar-evm
     - voteindexer
     - axelar-amplifier-verifier
+    - axelar-vald-heartbeats
 band-laozi-testnet6:
   chain_name: band
   protocol_type: cosmos
