+++ internal/common/packager.go
@@ -33,7 +33,7 @@ var (
 		"block", "balance", "upgrade", "block-data-analytics",
 
 		// Duties
-		"axelar-evm", "eventnonce", "oracle", "yoda",
+		"axelar-vald-heartbeats", "axelar-evm", "eventnonce", "oracle", "yoda",
 
 		// Babylon
 		"babylon-finality-provider-uptime", "babylon-btc-delegations",
