+++ internal/packages/duty/axelar-evm/parser/parser.go
@@ -3,9 +3,7 @@ package parser
 import (
 	"encoding/base64"
 	"encoding/json"
-	"fmt"
 
-	commontypes "github.com/cosmostation/cvms/internal/common/types"
 	"github.com/cosmostation/cvms/internal/packages/duty/axelar-evm/types"
 )
 
@@ -44,36 +42,4 @@ func AxelarProxyResisterParser(resp []byte) (types.AxelarProxyResisterStatus, er
 	}
 
 	return proxyResister, nil
-}
-
-func AxelarHeartbeatsFilterInTx(tx commontypes.CosmosTx, operatorAddr string) (bool, error) {
-	for _, rawMessage := range tx.Body.Messages {
-		var msg map[string]json.RawMessage
-		if err := json.Unmarshal(rawMessage, &msg); err != nil {
-			return false, fmt.Errorf("json unmashal error: %s", err)
-		}
-
-		if rawType, ok := msg["@type"]; ok {
-			var msgType string
-			if err := json.Unmarshal(rawType, &msgType); err != nil {
-				return false, fmt.Errorf("json unmashal error: %s", err)
-			}
-
-			if msgType == "/axelar.reward.v1beta1.RefundMsgRequest" {
-				if rawInnerMessage, ok := msg["inner_message"]; ok {
-					var innerMessage map[string]interface{}
-					if err := json.Unmarshal(rawInnerMessage, &innerMessage); err != nil {
-						return false, fmt.Errorf("json unmashal error: %s", err)
-					}
-
-					if innerMessage["@type"] == "/axelar.tss.v1beta1.HeartBeatRequest" {
-						if innerMessage["sender"] == operatorAddr {
-							return true, nil
-						}
-					}
-				}
-			}
-		}
-	}
-	return false, nil
 }
